Crie um programa que contenha um array de inteiros contendo 5 elementos. Utilizando apenas aritmética de ponteiros, leia esse array do teclado e imprima o dobro de cada valor lido.

+++ Lista1/questao11.c
@@ -10,15 +10,15 @@ Exercício: 11
 
 int main()
 {
-    float v[5];
+    int v[5];
         for(int i = 0; i < 5; i++){
             printf("\nDigite o valor da posicao [%d]: ", i + 1);
-            scanf("%f", &v[i]);
+            scanf("%d", &v[i]);
         }
         for(int i = 0; i < 5; i++){
-            float *x = &v[i];
+            int *x = &v[i];
             *x *= 2;
-            printf("Valor da posicao [%d] (dobro): %.2f\n", i + 1, *x);
+            printf("Valor da posicao [%d] (dobro): %d\n", i + 1, *x);
         }
 
         return 0;
